Rewrite Run simulation from to work with DialogService
